asks for species, gives graph w/ times
new additions: after BirdNET analyzes, script gives a list of the most
probable species. It asks the user for an input of which bird they 
want to investigate further. Then it gives a graph of that bird species
count for each time/audio using the time from the metadata comment.