performs simple audio analyzation on random audio
picks a random audio file from a given folder and provides:
the sampling rate, the spectrogram of the entire audio,
the highest audible frequency, the Fast Fourier Transform of
Entire Audio, duration of the audio. It also prompts the user
to input a time frame within the audio and in turn it provides:
a Spectrogram of the Segmented Time, the most prominent frequency,
graph of the Strongest Frequency per bin, and a graph of the
Power of Strongest Frequency per Bin.